api/requirements.txt
fastapi>=0.115.2
uvicorn==0.30.1
aiosqlite==0.20.0
python-dotenv==1.0.1
openpyxl==3.1.2
